rename mod hash to checksum
