+++ tests/func/test_stage.py
@@ -80,11 +80,11 @@ class TestReload(TestDvc):
 
         d = load_stage_file(stage.relpath)
 
-        from dvc.utils import hash
+        from dvc.utils import checksum as modchecksum
 
         # NOTE: checking that reloaded stage didn't change its checksum
         md5 = "11111111111111111111111111111111"
-        d[hash.CHECKSUM_MD5] = md5
+        d[modchecksum.CHECKSUM_MD5] = md5
         dump_stage_file(stage.relpath, d)
 
         stage = Stage.load(self.dvc, stage.relpath)
@@ -92,7 +92,7 @@ class TestReload(TestDvc):
         stage.dump()
 
         d = load_stage_file(stage.relpath)
-        self.assertEqual(d[hash.CHECKSUM_MD5], md5)
+        self.assertEqual(d[modchecksum.CHECKSUM_MD5], md5)
 
 
 class TestDefaultWorkingDirectory(TestDvc):
