+++ tests/func/test_state.py
@@ -4,7 +4,7 @@ import mock
 from dvc.utils.compat import str
 
 from dvc.state import State
-from dvc.utils.hash import file_checksum
+from dvc.utils.checksum import file_checksum
 from dvc.main import main
 from dvc.utils.fs import get_inode
