+++ tests/func/test_utils.py
@@ -3,7 +3,7 @@ import shutil
 import filecmp
 
 from dvc import utils
-from dvc.utils import hash
+from dvc.utils import checksum as modchecksum
 from tests.basic_env import TestDvc
 
 
@@ -37,7 +37,8 @@ class TestUtils(TestDvc):
             fd.write(b"a\r\nb\r\nc")
 
         self.assertEqual(
-            hash.file_checksum("cr")[0], hash.file_checksum("crlf")[0]
+            modchecksum.file_checksum("cr")[0],
+            modchecksum.file_checksum("crlf")[0],
         )
 
     def test_dict_md5(self):
@@ -61,7 +62,7 @@ class TestUtils(TestDvc):
         md5 = "8b263fa05ede6c3145c164829be694b4"
 
         self.assertEqual(
-            md5, hash.dict_checksum(d, exclude=["metric", "locked"])
+            md5, modchecksum.dict_checksum(d, exclude=["metric", "locked"])
         )
 
     def test_boxify(self):
